# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple blog service based on golang,mariadb and okteto",
        "title": "blog service",
        "contact": {
            "name": "Elchn"
        },
        "version": "v0.0"
    },
    "host": "https://blog-service-poiyyort.cloud.okteto.net",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/articletags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articletag"
                ],
                "summary": "get a list of articletags",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page index",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "size per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeeded",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleTagList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articletag"
                ],
                "summary": "create a new articletag",
                "parameters": [
                    {
                        "description": "article id",
                        "name": "article_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "tag id",
                        "name": "tag_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "maxLength": 100,
                        "minLength": 3,
                        "description": "creator",
                        "name": "created_by",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeeded",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleTag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/articletags/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articletag"
                ],
                "summary": "get an articletag with its id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "articletag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeeded",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleTag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articletag"
                ],
                "summary": "update an articletag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "articletag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "article id",
                        "name": "article_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "tag id",
                        "name": "tag_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "maxLength": 100,
                        "minLength": 3,
                        "description": "creator",
                        "name": "updated_by",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeeded",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleTag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articletag"
                ],
                "summary": "delete an articletag with given title or id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "articletag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeeded",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ArticleTag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "request errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal errors",
                        "schema": {
                            "$ref": "#/definitions/app.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "errors": {
                    "$ref": "#/definitions/errors.MyError"
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                }
            }
        },
        "app.Pager": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "app.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "data": {
                    "description": "Reference returns the reference document which maybe useful to solve this error.",
                    "type": "any"
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                }
            }
        },
        "errors.MyError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errors.MyError"
                    }
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                },
                "specific": {
                    "type": "string"
                }
            }
        },
...
        "model.ArticleTag": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_del": {
                    "type": "integer"
                },
                "tag_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "integer"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "model.ArticleTagList": {
            "type": "object",
            "properties": {
                "article_tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ArticleTag"
                    }
                },
                "pager": {
                    "$ref": "#/definitions/app.Pager"
                }
            }
        },
...
    }
}